requirements.txt
pandas==2.2.3
scikit-learn==1.5.2
scikit-surprise==1.1.4
numpy==1.26.4
tqdm==4.66.5
requests==2.32.3
pillow==10.4.0
customtkinter==5.2.2
ttkbootstrap==1.10.1
